+++ xxGraphicD3D6.cpp
@@ -453,7 +453,7 @@ uint64_t xxCreateVertexAttributeD3D6(uint64_t device, int count, int* attribute)
 
     d3dVertexAttribute.stride = stride;
 
-    return d3dVertexAttribute.value;
+    return static_cast<uint64_t>(d3dVertexAttribute.value);
 }
 //------------------------------------------------------------------------------
 void xxDestroyVertexAttributeD3D6(uint64_t vertexAttribute)
@@ -689,7 +689,7 @@ uint64_t xxCreateSamplerD3D6(uint64_t device, bool clampU, bool clampV, bool cla
         d3dSampler.minFilter = linearMin ? D3DTFN_ANISOTROPIC : D3DTFN_POINT;
     }
 
-    return d3dSampler.value;
+    return static_cast<uint64_t>(d3dSampler.value);
 }
 //------------------------------------------------------------------------------
 void xxDestroySamplerD3D6(uint64_t sampler)
@@ -720,7 +720,7 @@ uint64_t xxCreateBlendStateD3D6(uint64_t device, bool blending)
 {
     D3DRENDERSTATE3 d3dRenderState = {};
     d3dRenderState.alphaBlending = blending;
-    return d3dRenderState.value;
+    return static_cast<uint64_t>(d3dRenderState.value);
 }
 //------------------------------------------------------------------------------
 uint64_t xxCreateDepthStencilStateD3D6(uint64_t device, bool depthTest, bool depthWrite)
@@ -728,7 +728,7 @@ uint64_t xxCreateDepthStencilStateD3D6(uint64_t device, bool depthTest, bool dep
     D3DRENDERSTATE3 d3dRenderState = {};
     d3dRenderState.depthTest = depthTest;
     d3dRenderState.depthWrite = depthWrite;
-    return d3dRenderState.value;
+    return static_cast<uint64_t>(d3dRenderState.value);
 }
 //------------------------------------------------------------------------------
 uint64_t xxCreateRasterizerStateD3D6(uint64_t device, bool cull, bool scissor)
@@ -736,7 +736,7 @@ uint64_t xxCreateRasterizerStateD3D6(uint64_t device, bool cull, bool scissor)
     D3DRENDERSTATE3 d3dRenderState = {};
     d3dRenderState.cull = cull;
     d3dRenderState.scissor = scissor;
-    return d3dRenderState.value;
+    return static_cast<uint64_t>(d3dRenderState.value);
 }
 //------------------------------------------------------------------------------
 uint64_t xxCreatePipelineD3D6(uint64_t device, uint64_t renderPass, uint64_t blendState, uint64_t depthStencilState, uint64_t rasterizerState, uint64_t vertexAttribute, uint64_t vertexShader, uint64_t fragmentShader)
